src/api/requirements.txt
fastapi[standard]==0.115.12
pydantic==2.11.3
pydantic-settings==2.8.1
semantic-kernel[azure]==1.31.0
requests==2.32.3
azure-monitor-opentelemetry-exporter==1.0.0b36
python-multipart==0.0.20
azure-identity==1.23.0
async-lru==2.0.5